requirements.txt
openai>=1.30.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0